Use content_for in prev_next_pager, draw_interest_icons. Remove from @tabsets
