+++ app/views/admin/test.html.erb
@@ -8,8 +8,8 @@
   <%
     @title = "This is a test page, let's give it a much longer title just to see what happens when we do something like this"
 
+    draw_interest_icons(@user, Observation.first)
     tabs = create_tabs(admin_test_links)
-    tabs << draw_interest_icons(Observation.first)
     @tabsets = { right: draw_tab_set(tabs) }
   %>
