+++ app/views/application/content/_title_and_tab_sets.html.erb
@@ -8,21 +8,17 @@
 
 <div class="row">
   <!-- Push down pager, title on small, xs screens so buttons do not block -->
-  <%= content_tag(:div, safe_nbsp,
-                  class: "hidden-print visible-xs visible-sm mt-4") %>
+  <%= tag.div(safe_nbsp, class: "hidden-print visible-xs visible-sm mt-4") %>
 
   <div class="col-xs-<%= @tabsets[:right] ? 8 : 12 %>" id="title_bar">
     <!-- Pager -->
-    <%= content_tag(:div, class: "hidden-print") do
-          if @tabsets[:pager_for]
-            render(partial: "application/content/prev_next_pager",
-                  locals: { object: @tabsets[:pager_for] })
-          end
+    <%= tag.div(class: "hidden-print") do
+          yield :prev_next_pager
         end %>
 
     <!-- Title Caption -->
     <!-- e.g. "Observation 5"; Can be multiple lines, e.g., with Observer ID -->
-    <%= content_tag(:h1, @title, class: "h3", id: "title") -%>
+    <%= tag.h1(@title, class: "h3", id: "title") -%>
     <%= if @any_content_filters_applied
           add_context_help(content_tag(:span, "(#{:filtered.t})",
                                       class: "context-help"),
@@ -32,9 +28,12 @@
   </div>
 
   <!-- Tabsets -->
-  <%= content_tag_if(@tabsets[:right], :div, @tabsets[:right],
+  <%= content_tag_if(@tabsets[:right], :div,
                     id: "right_tabs",
-                    class: "hidden-print text-right col-sm-4 mb-3") %>
+                    class: "hidden-print text-right col-sm-4 mb-3") do
+        yield :interest_icons
+        @tabsets[:right]
+      end %>
 
   <%= content_tag_if(@full_width_tab_set, :div, @full_width_tab_set,
                     class: "hidden-print col-xs-12") %>
