+++ app/views/images/show.html.erb
@@ -1,7 +1,8 @@
 <%
 @title = :image_show_title.t(name: @image.unique_format_name)
+add_pager_for(@image)
 tabs = create_tabs(show_image_links(image: @image))
-@tabsets = { pager_for: @image, right: draw_tab_set(tabs) }
+@tabsets = { right: draw_tab_set(tabs) }
 @container = :wide
 %>
