+++ app/views/locations/descriptions/show.html.erb
@@ -1,8 +1,8 @@
 <%
 @title ||= @description.format_name.t
-
+add_pager_for(@description)
 tabs = create_tabs(show_description_links(description: @description))
-tabsets = { right: draw_tab_set(tabs), pager_for: @description }
+tabsets = { right: draw_tab_set(tabs) }
 @container = :wide
 %>
