+++ app/views/observations/show.html.erb
@@ -1,19 +1,20 @@
 <%
-  @owner_naming = owner_naming_line(@observation) # must before show_obs_title
-  @title = show_obs_title(obs: @observation, owner_naming: @owner_naming)
+@owner_naming = owner_naming_line(@observation) # must before show_obs_title
+@title = show_obs_title(obs: @observation, owner_naming: @owner_naming)
 
-  tabs = create_tabs(show_observation_links(obs: @observation, user: @user,
-                                            mappable: @mappable))
-  tabs << draw_interest_icons(@observation)
-  @tabsets = { pager_for: @observation, right: draw_tab_set(tabs) }
+add_pager_for(@observation)
+draw_interest_icons(@user, @observation)
+tabs = create_tabs(show_observation_links(obs: @observation, user: @user,
+                                          mappable: @mappable))
+@tabsets = { right: draw_tab_set(tabs) }
 
-  @container = :double
+@container = :double
 
-  show_map   = @user ? @user.thumbnail_maps : !session[:hide_thumbnail_maps]
-  show_lists = @user && @observation.species_lists.any?
-  show_links = @user && (@observation.external_links.any? || @new_sites.any?)
+show_map   = @user ? @user.thumbnail_maps : !session[:hide_thumbnail_maps]
+show_lists = @user && @observation.species_lists.any?
+show_links = @user && (@observation.external_links.any? || @new_sites.any?)
 
-  obs_locals = { observation: @observation }
+obs_locals = { observation: @observation }
 %>
 
 <div class="row">
