+++ app/views/projects/show.html.erb
@@ -1,12 +1,9 @@
 <%
 @title ||= :show_project_title.t(title: @project.title)
-
+add_pager_for(@project)
+draw_interest_icons(@user, @project)
 tabs = create_tabs(project_show_links(project: @project, user: @user))
-tabs << draw_interest_icons(@project)
-@tabsets = {
-  pager_for: @project,
-  right: draw_tab_set(tabs)
-}
+@tabsets = { right: draw_tab_set(tabs) }
 %>
 
 <div class="row">
