+++ app/views/rss_logs/show.html.erb
@@ -1,6 +1,6 @@
 <%
-  @title = :show_rss_log_title.t(title: @rss_log.unique_format_name)
-  @tabsets = { pager_for: @rss_log }
+@title = :show_rss_log_title.t(title: @rss_log.unique_format_name)
+add_pager_for(@rss_log)
 %>
 
 <% if @rss_log.notes %>
