+++ app/views/species_lists/show.html.erb
@@ -1,11 +1,11 @@
 <%
-  @title = :species_list_show_title.t(name: @species_list.unique_format_name)
-
-  links = species_list_show_links(list: @species_list)
-  tabs = create_tabs(links)
-  tabs << draw_interest_icons(@species_list) if @user
-  @tabsets = { pager_for: @species_list, right: draw_tab_set(tabs) }
-  @container = :text_image
+@title = :species_list_show_title.t(name: @species_list.unique_format_name)
+add_pager_for(@species_list)
+draw_interest_icons(@user, @species_list)
+links = species_list_show_links(list: @species_list)
+tabs = create_tabs(links)
+@tabsets = { right: draw_tab_set(tabs) }
+@container = :text_image
 %>
 
 <div class="container-text">
