+++ app/views/users/show.html.erb
@@ -1,16 +1,18 @@
 <%
-  id = @show_user.id
-  name = @show_user.unique_text_name
-  @title = :show_user_title.t(user: name)
-  @container = :full
+id = @show_user.id
+name = @show_user.unique_text_name
+@title = :show_user_title.t(user: name)
+add_pager_for(@show_user)
 
-  links = user_show_links(show_user: @show_user, user: @user)
-  tabs = create_tabs(links)
-  @tabsets = { pager_for: @show_user, right: draw_tab_set(tabs) }
+links = user_show_links(show_user: @show_user, user: @user)
+tabs = create_tabs(links)
+@tabsets = { right: draw_tab_set(tabs) }
 
-  rows = user_stats_rows(@show_user, @user_data)
-  paths = user_stats_link_paths(@show_user)
-  total = 0
+@container = :full
+
+rows = user_stats_rows(@show_user, @user_data)
+paths = user_stats_link_paths(@show_user)
+total = 0
 %>
 
 <div class="row">
